AV - Add a remove function to the inventory class; also add a missing test for the constructor of GroceryItem (whoopsies)

+++ src/main/groovy/avild/grocerykata/GroceryItem.groovy
@@ -4,9 +4,9 @@ class GroceryItem {
     String name
     int price
     int markdownPrice
-    long weight
+    float weight
 
-    GroceryItem(String name, int price, int markdownPrice, long weight) {
+    GroceryItem(String name, int price, int markdownPrice, float weight) {
         this.name = name
         this.price = price
         this.markdownPrice = markdownPrice
